test: group dockerfile tests
